[CL-677] Move logic away from builder frame
