+++ front/app/modules/commercial/content_builder/admin/components/ContentBuilderFrame/ContentBuilderFrame.test.tsx
@@ -1,37 +1,21 @@
 import React from 'react';
 import { render } from 'utils/testUtils/rtl';
 import ContentBuilderFrame from './';
-import { Editor } from '@craftjs/core';
-import { IContentBuilderLayoutData } from '../../../services/contentBuilder';
+import { Editor, SerializedNodes } from '@craftjs/core';
 
-let mockEditorData: IContentBuilderLayoutData = {
-  id: '2',
-  type: 'content_builder_layout',
-  attributes: {
-    craftjs_jsonmultiloc: {
-      en: {
-        nodeId: {
-          custom: {},
-          displayName: 'div',
-          hidden: false,
-          isCanvas: true,
-          linkedNodes: {},
-          nodes: [],
-          type: 'div',
-          props: {},
-          parent: 'ROOT',
-        },
-      },
-    },
-    code: 'project_description',
-    enabled: true,
+const mockEditorData: SerializedNodes = {
+  nodeId: {
+    custom: {},
+    displayName: 'div',
+    hidden: false,
+    isCanvas: true,
+    linkedNodes: {},
+    nodes: [],
+    type: 'div',
+    props: {},
+    parent: 'ROOT',
   },
 };
-
-jest.mock('hooks/useLocale');
-jest.mock('../../../hooks/useContentBuilder', () => {
-  return jest.fn(() => ({ data: mockEditorData }));
-});
 
 const mockDeserialize = jest.fn();
 
@@ -47,26 +31,15 @@ describe('ContentBuilderFrame', () => {
   it('calls deserialize correctly when there is data', () => {
     render(
       <Editor>
-        <ContentBuilderFrame projectId="id" />
+        <ContentBuilderFrame editorData={mockEditorData} />
       </Editor>
     );
-    expect(mockDeserialize).toHaveBeenCalledWith(
-      mockEditorData.attributes.craftjs_jsonmultiloc.en
-    );
+    expect(mockDeserialize).toHaveBeenCalledWith(mockEditorData);
   });
   it('does not call deserialize when there is no data', () => {
-    mockEditorData = {
-      id: '2',
-      type: 'content_builder_layout',
-      attributes: {
-        craftjs_jsonmultiloc: {},
-        code: 'project_description',
-        enabled: true,
-      },
-    };
     render(
       <Editor>
-        <ContentBuilderFrame projectId="id" />
+        <ContentBuilderFrame editorData={undefined} />
       </Editor>
     );
     expect(mockDeserialize).not.toHaveBeenCalled();
