+++ front/app/modules/commercial/content_builder/admin/components/ContentBuilderFrame/index.tsx
@@ -1,31 +1,18 @@
 import React, { useEffect } from 'react';
-import { Frame, Element, useEditor } from '@craftjs/core';
+import { Frame, Element, useEditor, SerializedNode } from '@craftjs/core';
 
-// utils
-import { isNilOrError } from 'utils/helperUtils';
+type ContentBuilderFrame = {
+  editorData?: Record<string, SerializedNode>;
+};
 
-// hooks
-import useLocale from 'hooks/useLocale';
-import useContentBuilderLayout from '../../../hooks/useContentBuilder';
-import { PROJECT_DESCRIPTION_CODE } from '../../../services/contentBuilder';
-
-const ContentBuilderFrame = ({ projectId }: { projectId: string }) => {
+const ContentBuilderFrame = ({ editorData }: ContentBuilderFrame) => {
   const { actions } = useEditor();
-  const data = useContentBuilderLayout({
-    projectId,
-    code: PROJECT_DESCRIPTION_CODE,
-  });
-  const locale = useLocale();
 
   useEffect(() => {
-    const editorData =
-      !isNilOrError(data) && !isNilOrError(locale)
-        ? data.data.attributes.craftjs_jsonmultiloc[locale]
-        : undefined;
     if (editorData) {
       actions.deserialize(editorData);
     }
-  }, [actions, data, locale]);
+  }, [editorData, actions]);
 
   return (
     <Frame>
