+++ front/app/modules/commercial/content_builder/admin/components/ContentBuilderToolbox/index.tsx
@@ -26,6 +26,7 @@ import messages from '../../messages';
 
 // Styles
 import styled from 'styled-components';
+import { colors } from 'utils/styleUtils';
 
 const DraggableElement = styled.div`
   cursor: move;
@@ -41,119 +42,143 @@ const ContentBuilderToolbox = ({
   } = useEditor();
 
   return (
-    <Box w="100%" display="inline">
-      <DraggableElement
-        id="e2e-draggable-single-column"
-        ref={(ref) =>
-          ref &&
-          connectors.create(
-            ref,
-            <Element canvas is={Container} id="container" />
-          )
-        }
-      >
-        <ToolboxItem icon="column1" label={formatMessage(messages.oneColumn)} />
-      </DraggableElement>
-      <DraggableElement
-        id="e2e-draggable-two-column"
-        ref={(ref) =>
-          ref &&
-          connectors.create(
-            ref,
-            <Element canvas is={TwoColumn} columnLayout="1-1" id="twoColumn" />
-          )
-        }
-      >
-        <ToolboxItem icon="column2" label={formatMessage(messages.twoColumn)} />
-      </DraggableElement>
-      <DraggableElement
-        id="e2e-draggable-three-column"
-        ref={(ref) =>
-          ref &&
-          connectors.create(
-            ref,
-            <Element canvas is={ThreeColumn} id="threeColumn" />
-          )
-        }
-      >
-        <ToolboxItem
-          icon="column3"
-          label={formatMessage(messages.threeColumn)}
-        />
-      </DraggableElement>
-      <DraggableElement
-        id="e2e-draggable-text"
-        ref={(ref) =>
-          ref &&
-          connectors.create(
-            ref,
-            <Element
-              is={Text}
-              id="text"
-              text={formatMessage(messages.textValue)}
-            />,
-            {
-              onCreate: (node) => {
-                selectNode(node.rootNodeId);
-              },
-            }
-          )
-        }
-      >
-        <ToolboxItem icon="text" label={formatMessage(messages.text)} />
-      </DraggableElement>
-      <DraggableElement
-        id="e2e-draggable-image"
-        ref={(ref) => {
-          ref &&
+    <Box
+      position="fixed"
+      zIndex="3"
+      flex="0 0 auto"
+      h="100%"
+      w="210px"
+      display="flex"
+      flexDirection="column"
+      alignItems="center"
+      bgColor="#ffffff"
+      borderRight={`1px solid ${colors.mediumGrey}`}
+    >
+      <Box w="100%" display="inline">
+        <DraggableElement
+          id="e2e-draggable-single-column"
+          ref={(ref) =>
+            ref &&
             connectors.create(
               ref,
-              <Element is={Image} imageUrl="" id="image" alt="" />,
+              <Element canvas is={Container} id="container" />
+            )
+          }
+        >
+          <ToolboxItem
+            icon="column1"
+            label={formatMessage(messages.oneColumn)}
+          />
+        </DraggableElement>
+        <DraggableElement
+          id="e2e-draggable-two-column"
+          ref={(ref) =>
+            ref &&
+            connectors.create(
+              ref,
+              <Element
+                canvas
+                is={TwoColumn}
+                columnLayout="1-1"
+                id="twoColumn"
+              />
+            )
+          }
+        >
+          <ToolboxItem
+            icon="column2"
+            label={formatMessage(messages.twoColumn)}
+          />
+        </DraggableElement>
+        <DraggableElement
+          id="e2e-draggable-three-column"
+          ref={(ref) =>
+            ref &&
+            connectors.create(
+              ref,
+              <Element canvas is={ThreeColumn} id="threeColumn" />
+            )
+          }
+        >
+          <ToolboxItem
+            icon="column3"
+            label={formatMessage(messages.threeColumn)}
+          />
+        </DraggableElement>
+        <DraggableElement
+          id="e2e-draggable-text"
+          ref={(ref) =>
+            ref &&
+            connectors.create(
+              ref,
+              <Element
+                is={Text}
+                id="text"
+                text={formatMessage(messages.textValue)}
+              />,
               {
                 onCreate: (node) => {
                   selectNode(node.rootNodeId);
                 },
               }
-            );
-        }}
-      >
-        <ToolboxItem icon="image" label={formatMessage(messages.image)} />
-      </DraggableElement>
-      <DraggableElement
-        id="e2e-draggable-iframe"
-        ref={(ref) =>
-          ref &&
-          connectors.create(
-            ref,
-            <Element
-              is={Iframe}
-              id="Iframe"
-              url=""
-              height={500}
-              hasError={false}
-            />,
-            {
-              onCreate: (node) => {
-                selectNode(node.rootNodeId);
-              },
-            }
-          )
-        }
-      >
-        <ToolboxItem icon="code" label={formatMessage(messages.url)} />
-      </DraggableElement>
-      <DraggableElement
-        id="e2e-draggable-about-box"
-        ref={(ref) =>
-          ref &&
-          connectors.create(
-            ref,
-            <Element is={AboutBox} id="AboutBox" projectId={projectId} />
-          )
-        }
-      >
-        <ToolboxItem icon="info3" label={formatMessage(messages.aboutBox)} />
-      </DraggableElement>
+            )
+          }
+        >
+          <ToolboxItem icon="text" label={formatMessage(messages.text)} />
+        </DraggableElement>
+        <DraggableElement
+          id="e2e-draggable-image"
+          ref={(ref) => {
+            ref &&
+              connectors.create(
+                ref,
+                <Element is={Image} imageUrl="" id="image" alt="" />,
+                {
+                  onCreate: (node) => {
+                    selectNode(node.rootNodeId);
+                  },
+                }
+              );
+          }}
+        >
+          <ToolboxItem icon="image" label={formatMessage(messages.image)} />
+        </DraggableElement>
+        <DraggableElement
+          id="e2e-draggable-iframe"
+          ref={(ref) =>
+            ref &&
+            connectors.create(
+              ref,
+              <Element
+                is={Iframe}
+                id="Iframe"
+                url=""
+                height={500}
+                hasError={false}
+              />,
+              {
+                onCreate: (node) => {
+                  selectNode(node.rootNodeId);
+                },
+              }
+            )
+          }
+        >
+          <ToolboxItem icon="code" label={formatMessage(messages.url)} />
+        </DraggableElement>
+        <DraggableElement
+          id="e2e-draggable-about-box"
+          ref={(ref) =>
+            ref &&
+            connectors.create(
+              ref,
+              <Element is={AboutBox} id="AboutBox" projectId={projectId} />
+            )
+          }
+        >
+          <ToolboxItem icon="info3" label={formatMessage(messages.aboutBox)} />
+        </DraggableElement>
+      </Box>
     </Box>
   );
 };
