+++ front/app/modules/commercial/content_builder/admin/containers/index.tsx
@@ -3,7 +3,7 @@ import { withRouter } from 'react-router';
 
 // styles
 import styled from 'styled-components';
-import { colors, stylingConsts } from 'utils/styleUtils';
+import { stylingConsts } from 'utils/styleUtils';
 
 // components
 import { RightColumn } from 'containers/Admin';
@@ -15,6 +15,10 @@ import ContentBuilderToolbox from '../components/ContentBuilderToolbox';
 import ContentBuilderTopBar from '../components/ContentBuilderTopBar';
 import ContentBuilderFrame from '../components/ContentBuilderFrame';
 import ContentBuilderSettings from '../components/ContentBuilderSettings';
+import { PROJECT_DESCRIPTION_CODE } from '../../services/contentBuilder';
+import useLocale from 'hooks/useLocale';
+import useContentBuilderLayout from '../../hooks/useContentBuilder';
+import { isNilOrError } from 'utils/helperUtils';
 
 const StyledRightColumn = styled(RightColumn)`
   min-height: calc(100vh - ${2 * stylingConsts.menuHeight}px);
@@ -27,6 +31,18 @@ const StyledRightColumn = styled(RightColumn)`
 const ContentBuilderPage = ({ params: { projectId } }) => {
   const [mobilePreviewEnabled, setMobilePreviewEnabled] = useState(false);
 
+  const locale = useLocale();
+
+  const contentBuilderLayout = useContentBuilderLayout({
+    projectId,
+    code: PROJECT_DESCRIPTION_CODE,
+  });
+
+  const editorData =
+    !isNilOrError(contentBuilderLayout) && !isNilOrError(locale)
+      ? contentBuilderLayout.data.attributes.craftjs_jsonmultiloc[locale]
+      : undefined;
+
   return (
     <Box display="flex" flexDirection="column" w="100%">
       <Editor isPreview={false}>
@@ -35,23 +51,10 @@ const ContentBuilderPage = ({ params: { projectId } }) => {
           setMobilePreviewEnabled={setMobilePreviewEnabled}
         />
         <Box mt={`${stylingConsts.menuHeight}px`} display="flex">
-          <Box
-            position="fixed"
-            zIndex="3"
-            flex="0 0 auto"
-            h="100%"
-            w="210px"
-            display="flex"
-            flexDirection="column"
-            alignItems="center"
-            bgColor="#ffffff"
-            borderRight={`1px solid ${colors.mediumGrey}`}
-          >
-            <ContentBuilderToolbox />
-          </Box>
+          <ContentBuilderToolbox />
           <StyledRightColumn>
             <Box width="1000px">
-              <ContentBuilderFrame projectId={projectId} />
+              <ContentBuilderFrame editorData={editorData} />
             </Box>
           </StyledRightColumn>
           <ContentBuilderSettings />
